add option for returning preds and labels in evaluator. avoid duplicate execution of test eval + model loading. cleaner logging.
