+++ examples/doc_classification_xval.py
@@ -1,10 +1,10 @@
 # fmt: off
 import logging
+import json
 
-from farm.data_handler.data_silo import DataSilo
+from farm.data_handler.data_silo import DataSilo, DataSilo4Xval
 from farm.data_handler.processor import TextClassificationProcessor
 from farm.modeling.optimization import initialize_optimizer
-from farm.infer import Inferencer
 from farm.modeling.adaptive_model import AdaptiveModel
 from farm.modeling.language_model import LanguageModel
 from farm.modeling.prediction_head import TextClassificationHead
@@ -14,18 +14,17 @@ from farm.utils import set_all_seeds, MLFlowLogger, initialize_device_settings
 from farm.eval import Evaluator
 from sklearn.metrics import matthews_corrcoef, recall_score, precision_score, f1_score, mean_squared_error, r2_score
 from farm.metrics import simple_accuracy, register_metrics
-from torch.utils.data import ConcatDataset, Subset
-import torch
-from sklearn.model_selection import StratifiedKFold, KFold
-from torch.utils.data.distributed import DistributedSampler
-from torch.utils.data.sampler import RandomSampler, SequentialSampler
 
-from farm.data_handler.dataloader import NamedDataLoader
-
+##########################
+########## Logging
+##########################
+logger = logging.getLogger(__name__)
 logging.basicConfig(
     format="%(asctime)s - %(levelname)s - %(name)s -   %(message)s",
     datefmt="%m/%d/%Y %H:%M:%S",
     level=logging.INFO)
+# reduce verbosity from transformers library
+logging.getLogger('transformers').setLevel(logging.WARNING)
 
 # ml_logger = MLFlowLogger(tracking_uri="https://public-mlflow.deepset.ai/")
 # for local logging instead:
@@ -35,12 +34,12 @@ ml_logger = MLFlowLogger(tracking_uri="logs")
 ##########################
 ########## Settings
 ##########################
-xval_folds = 10
+xval_folds = 3 #10
 xval_stratified = True
 
 set_all_seeds(seed=42)
 device, n_gpu = initialize_device_settings(use_cuda=True)
-n_epochs = 20
+n_epochs = 1
 batch_size = 32
 evaluate_every = 100
 lang_model = "bert-base-german-cased"
@@ -49,12 +48,6 @@ lang_model = "bert-base-german-cased"
 tokenizer = Tokenizer.load(
     pretrained_model_name_or_path=lang_model,
     do_lower_case=False)
-
-# 2. Create a DataProcessor that handles all the conversion from raw text into a pytorch Dataset
-# Here we load GermEval 2018 Data.
-
-# The processor wants to know the possible labels ...
-label_list = ["OTHER", "OFFENSE"]
 
 # The evaluation on the dev-set can be done with one of the predefined metrics or with a
 # metric defined as a function from (preds, labels) to a dict that contains all the actual
@@ -75,13 +68,16 @@ def mymetrics(preds, labels):
         "f1_offense": f1offense,
         "f1_macro": f1macro,
         "f1_micro": f1micro,
-        "mcc": mcc,
-        "_preds": preds,
-        "_labels": labels
+        "mcc": mcc
     }
 register_metrics('mymetrics', mymetrics)
 metric = 'mymetrics'
 
+# 2. Create a DataProcessor that handles all the conversion from raw text into a pytorch Dataset
+# Here we load GermEval 2018 Data.
+
+# The processor wants to know the possible labels ...
+label_list = ["OTHER", "OFFENSE"]
 processor = TextClassificationProcessor(tokenizer=tokenizer,
                                         max_seq_len=64,
                                         data_dir="../data/germeval18",
@@ -95,98 +91,13 @@ data_silo = DataSilo(
     processor=processor,
     batch_size=batch_size)
 
-# For performing cross validation, we really want to combine all the instances from all
-# the sets or just some of the sets, then create a different data silo instance for each fold.
-# Here, we combine the instances from the train and dev sets to perform xcross validation,
-# then create a different data silo instance with train, dev and test sets for each fold
-# We use our own DataSiloTmp class to just represent the subsets for train/dev/test for each fold
-# as we need it
-class DataSilo4Xval:
-    def __init__(self, origsilo, trainset, devset, testset):
-        self.tensor_names = origsilo.tensor_names
-        self.data = {"train":trainset, "dev":devset, "test":testset}
-        self.processor = origsilo.processor
-        self.batch_size = origsilo.batch_size
-        # should not be necessary, xval makes no sense with huge data
-        # sampler_train = DistributedSampler(self.data["train"])
-        sampler_train = RandomSampler(trainset)
-
-        self.data_loader_train = NamedDataLoader(
-            dataset=trainset,
-            sampler=sampler_train,
-            batch_size=self.batch_size,
-            tensor_names=self.tensor_names,
-        )
-        self.data_loader_dev = NamedDataLoader(
-            dataset=devset,
-            sampler=SequentialSampler(devset),
-            batch_size=self.batch_size,
-            tensor_names=self.tensor_names,
-        )
-        self.data_loader_test = NamedDataLoader(
-            dataset=testset,
-            sampler=SequentialSampler(testset),
-            batch_size=self.batch_size,
-            tensor_names=self.tensor_names,
-        )
-        self.loaders = {
-            "train": self.data_loader_train,
-            "dev": self.data_loader_dev,
-            "test": self.data_loader_test,
-        }
-
-    def get_data_loader(self, which):
-        return self.loaders[which]
-
-    @staticmethod
-    def make(datasilo, sets=["train", "dev", "test"], n_splits=5, stratified=True,
-             shuffle=True, random_state=None, dev_split=0.2):
-        """
-        Create number of folds data-silo-like objects which can be used for training from the
-        original data silo passed on.
-        :param datasilo: the data silo that contains the original data
-        :param sets: which sets to use to create the xval folds
-        :param n_splits: number of folds to create
-        :param stratified: if class stratificiation should be done
-        :param shuffle: shuffle each class' samples before splitting
-        :param random_state: random state for shuffling
-        :param dev_split: size of the dev set for a fold, held out from the training set
-        """
-        setstoconcat = [datasilo.data[setname] for setname in sets]
-        ds_all = ConcatDataset(setstoconcat)
-        idxs = list(range(len(ds_all)))
-        if stratified:
-            # get all the labels for stratification
-            ytensors = [t[3][0] for t in ds_all]
-            Y = torch.stack(ytensors)
-            xval = StratifiedKFold(n_splits=n_splits, shuffle=shuffle, random_state=random_state)
-            xval_split = xval.split(idxs,Y)
-        else:
-            xval = KFold(n_splits=n_splits, shuffle=shuffle, random_state=random_state)
-            xval_split = xval.split(idxs)
-        # for each fold create a DataSilo4Xval instance, where the training set is further
-        # divided into actual train and dev set
-        silos = []
-        for train_idx, test_idx in xval_split:
-            n_dev = int(dev_split * len(train_idx))
-            n_actual_train = len(train_idx) - n_dev
-            # TODO: this split into actual train and test set could/should also be stratified, for now
-            # we just do this by taking the first/last indices from the train set (which should be
-            # shuffled by default)
-            actual_train_idx = train_idx[:n_actual_train]
-            dev_idx = train_idx[n_actual_train:]
-            # create the actual datasets
-            ds_train = Subset(ds_all, actual_train_idx)
-            ds_dev = Subset(ds_all, dev_idx)
-            ds_test = Subset(ds_all, test_idx)
-            silos.append(DataSilo4Xval(datasilo, ds_train, ds_dev, ds_test))
-        return silos
-
-silos = DataSilo4Xval.make(data_silo, n_splits=3)
+# Load one silo for each fold in our cross-validation
+silos = DataSilo4Xval.make(data_silo, n_splits=xval_folds)
 
 # the following steps should be run for each of the folds of the cross validation, so we put them
 # into a function
-def train_on_split(silo_to_use, foldnr):
+def train_on_split(silo_to_use, n_fold, save_dir):
+    logger.info(f"############ Crossvalidation: Fold {n_fold} ############")
     # Create an AdaptiveModel
     # a) which consists of a pretrained language model as a basis
     language_model = LanguageModel.load(lang_model)
@@ -215,13 +126,11 @@ def train_on_split(silo_to_use, foldnr):
 
     # An early stopping instance can be used to save the model that performs best on the dev set
     # according to some metric and stop training when no improvement is happening for some iterations.
-    # NOTE: if we would use a different save directory for each fold, we could afterwards use a the
+    # NOTE: Using a different save directory for each fold, allows us afterwards to use the
     # nfolds best models in an ensemble!
-    save_dir = "saved_models/bert-german-doc-tutorial-es-{}".format(foldnr)
+    save_dir += f"-{n_fold}"
     earlystopping = EarlyStopping(
         metric="f1_offense", mode="max",   # use the metric from our own metrics function instead of loss
-        # metric="f1_macro", mode="max",  # use f1_macro from the dev evaluator of the trainer
-        # metric="loss", mode="min",   # use loss from the dev evaluator of the trainer
         save_dir=save_dir,  # where to save the best model
         patience=5    # number of evaluations to wait for improvement before terminating the training
     )
@@ -234,52 +143,51 @@ def train_on_split(silo_to_use, foldnr):
         warmup_linear=warmup_linear,
         evaluate_every=evaluate_every,
         device=device,
-        early_stopping=earlystopping)
+        early_stopping=earlystopping,
+        evaluator_test=False)
 
     # train it
     model = trainer.train(model)
 
-    # Since we used early stopping, restore the best model from there.
-    lm_name = model.language_model.name
-    model = AdaptiveModel.load(earlystopping.save_dir, trainer.device, lm_name=lm_name)
-    model.connect_heads_with_processor(silo_to_use.processor.tasks, require_labels=True)
     return model
-
 
 # for each fold, run the whole training, earlystopping to get a model, then evaluate the model
 # on the test set of each fold
 # Remember all the results for overall metrics over all predictions of all folds and for averaging
 allresults = []
+all_preds = []
+all_labels = []
 bestfold = None
 bestf1_offense = -1
-for foldnr, silo in enumerate(silos):
-    model = train_on_split(silo, foldnr)
-    # make an evaluator for the evaluation on the test set
+save_dir = "saved_models/bert-german-doc-tutorial-es"
+for num_fold, silo in enumerate(silos):
+    model = train_on_split(silo, num_fold, save_dir)
+
+    # do eval on test set here (and not in Trainer),
+    #  so that we can easily store the actual preds and labels for a "global" eval across all folds.
     evaluator_test = Evaluator(
         data_loader=silo.get_data_loader("test"),
         tasks=silo.processor.tasks,
         device=device
     )
-    result = evaluator_test.eval(model)
+    result = evaluator_test.eval(model, return_preds_and_labels=True)
+    evaluator_test.log_results(result, "Test", steps=len(silo.get_data_loader("test")), num_fold=num_fold)
+
     allresults.append(result)
-    # get the f1_offense metric
+    all_preds.extend(result[0].get("preds"))
+    all_labels.extend(result[0].get("labels"))
+
+    # keep track of best fold
     f1_offense = result[0]["f1_offense"]
     if f1_offense > bestf1_offense:
         bestf1_offense = f1_offense
-        bestfold = foldnr
+        bestfold = num_fold
 
 # Save the per-fold results to json for a separate, more detailed analysis
-import json
 with open("doc_classification_xval.results.json", "wt") as fp:
     json.dump(allresults, fp)
 
-# Combine the per-fold labels/predictions so we can calculate overall metrics
-all_preds = []
-all_labels = []
-for result in allresults:
-    all_preds.extend(result[0].get("_preds"))
-    all_labels.extend(result[0].get("_labels"))
-
+# calculate overall metrics across all folds
 xval_f1_micro = f1_score(all_labels, all_preds, labels=label_list, average="micro")
 xval_f1_macro = f1_score(all_labels, all_preds, labels=label_list, average="macro")
 xval_f1_offense = f1_score(all_labels, all_preds, labels=label_list, pos_label="OFFENSE")
@@ -295,6 +203,7 @@ print("XVAL MCC:        ", xval_mcc)
 
 # Just for illustration, use the best model from the best xval val for evaluation on
 # the original (still unseen) test set.
+logger.info("###### Final Eval on hold out test set using best model #####")
 evaluator_origtest = Evaluator(
     data_loader=data_silo.get_data_loader("test"),
     tasks=data_silo.processor.tasks,
