+++ farm/eval.py
@@ -43,12 +43,14 @@ class Evaluator:
         #self.metrics = metrics
         self.classification_report = classification_report
 
-    def eval(self, model):
+    def eval(self, model, return_preds_and_labels=False):
         """
         Performs evaluation on a given model.
 
         :param model: The model on which to perform evaluation
         :type model: AdaptiveModel
+        :param return_preds_and_labels: Whether to add preds and labels in the returned dicts of the
+        :type return_preds_and_labels: bool
         :return all_results: A list of dictionaries, one for each prediction head. Each dictionary contains the metrics
                              and reports generated during evaluation.
         :rtype all_results: list of dicts
@@ -141,17 +143,24 @@ class Evaluator:
                         labels=all_possible_labels,
                         target_names=head.label_list)
 
+            if return_preds_and_labels:
+                result["preds"] = preds_all[head_num]
+                result["labels"] = label_all[head_num]
+
             all_results.append(result)
 
         return all_results
 
     @staticmethod
-    def log_results(results, dataset_name, steps, logging=True, print=True):
+    def log_results(results, dataset_name, steps, logging=True, print=True, num_fold=None):
         # Print a header
         header = "\n\n"
         header += BUSH_SEP + "\n"
         header += "***************************************************\n"
-        header += f"***** EVALUATION | {dataset_name.upper()} SET | AFTER {steps} BATCHES *****\n"
+        if num_fold:
+            header += f"***** EVALUATION | FOLD: {num_fold} | {dataset_name.upper()} SET | AFTER {steps} BATCHES *****\n"
+        else:
+            header += f"***** EVALUATION | {dataset_name.upper()} SET | AFTER {steps} BATCHES *****\n"
         header += "***************************************************\n"
         header += BUSH_SEP + "\n"
         logger.info(header)
@@ -161,7 +170,7 @@ class Evaluator:
             for metric_name, metric_val in head.items():
                 # log with ML framework (e.g. Mlflow)
                 if logging:
-                    if not metric_name.startswith("_"):
+                    if not metric_name in ["preds","labels"]:
                         if isinstance(metric_val, numbers.Number):
                             MlLogger.log_metrics(
                                 metrics={
@@ -176,5 +185,5 @@ class Evaluator:
                             metric_val = metric_val[:7500] + "\n ............................. \n" + metric_val[-500:]
                         logger.info("{}: \n {}".format(metric_name, metric_val))
                     else:
-                        if not metric_name.startswith("_"):
+                        if not metric_name in ["preds", "labels"]:
                             logger.info("{}: {}".format(metric_name, metric_val))
